move connect method behind interface
